backend/requirements.txt
fastapi
uvicorn
google.generativeai
qdrant-client
pypdf
python-dotenv
pydantic
pdfplumber
pytesseract
python-multipart
pdfminer.six
pdf2image